+++ docs/project/changelog.md
@@ -15,6 +15,9 @@ substitutions:
 - {{ API }} {any}`pyodide.runPythonAsync` no longer automatically calls
   {any}`pyodide.loadPackagesFromImports`.
   {pr}`1538`.
+- {{ Enhancement }} Added the {any}`PyProxy.callKwargs` method to allow using
+  Python functions with keyword arguments from Javascript.
+  {pr}`1539`
 
 ## Version 0.17.0
 *April 21, 2020*